move ApiTestSteps
