+++ src/tests/api/ApiUtils.ts
@@ -1,6 +1,6 @@
 import Ajv from "ajv";
 import { XMLParser } from "fast-xml-parser";
-import { HARNESS_API_INSTANCE } from "./backend/ApiTestSteps";
+import { HARNESS_API_INSTANCE } from "./ApiTestSteps";
 import { TxmaEvent, TxmaEventName } from "../../utils/TxmaEvent";
 import * as F2F_CRI_AUTH_CODE_ISSUED_SCHEMA from "../data/F2F_CRI_AUTH_CODE_ISSUED_SCHEMA.json";
 import * as F2F_CRI_END_SCHEMA from "../data/F2F_CRI_END_SCHEMA.json";
